remove nested modules
