--- Stanislav_Dzisiak/2/lib/url_analyzer/utils.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module UrlAnalyzer
-  module Utils
-    autoload 'UrlList', 'url_analyzer/utils/url_list'
-    autoload 'Parser', 'url_analyzer/utils/parser'
-  end
-end
